Relations Added relations to model classes. Added abstract non implemented method for input validation to models.

+++ database/migrations/2017_02_07_223806_create_sections_table.php
@@ -21,7 +21,7 @@ class CreateSectionsTable extends Migration
             $table->string('days', 7);
             $table->string('professor')->nullable()->default(null);
             $table->string('building')->nullable()->default(null);
-            $table->tinyInteger('room_number')->nullable()->default(null);
+            $table->smallInteger('room_number')->nullable()->default(null);
         });
 
         Schema::table('sections', function (Blueprint $table){
